HDF5 moved to its own .c file.
binary-binary and binary-hdf5 tests passing.
